FastApi-Course-Management/requirements.txt
fastapi[all]
fastapi
uvicorn
sqlalchemy
pydantic
passlib[bcrypt]
python-jose
pymongo
pymysql
databases
httpx